requirements.txt
# Core Django
Django==5.0.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# WaafiPay Integration
requests==2.31.0
waafipay-python==0.1.0  # WaafiPay SDK

# Forms & Validation
django-crispy-forms==2.3
crispy-bootstrap5==2.0.0

# File Handling
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34  # For S3 storage in production

# Security
django-cors-headers==4.3.1

# Utils
python-decouple==3.8
python-dotenv==1.0.1

# Development
django-debug-toolbar==4.2.0

# Testing
coverage==7.4.1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Celery for async tasks (optional)
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
